Refactor BookingData model to change amount fields from double to int and update booking details display logic

+++ lib/model/bookings/booking_details_model.dart
@@ -33,7 +33,7 @@ class BookingData {
   int? distance;
   int? noOfSeats;
   int? originalAmount;
-  double? totalAmount;
+  int? totalAmount;
   String? status;
   String? paymentStatus;
   Null cardId;
@@ -41,7 +41,7 @@ class BookingData {
   Null refundId;
   int? refundAmount;
   int? cancellationFee;
-  double? platformFee;
+  int? platformFee;
   int? discount;
   int? amountPayableToDriver;
   String? createdAt;
